// apps/web/src/generated/wire/AnswerSchema.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AnswerSchema = { "kind": "integer", min: string, max: string, } | { "kind": "rational", max_abs_numerator: number, max_denominator: number, require_reduced_fraction_form: boolean, } | { "kind": "decimal", max_scale: number, } | { "kind": "ordered_pair" } | { "kind": "ordered_tuple", length: number, } | { "kind": "algebraic" };