// apps/web/src/generated/wire/ArithmeticExpression.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ArithmeticOperator } from "./ArithmeticOperator";
import type { RationalCoefficient } from "./RationalCoefficient";

export type ArithmeticExpression = { "kind": "integer", value: number, } | { "kind": "rational", value: RationalCoefficient, } | { "kind": "exact_decimal", coefficient: number, scale: number, } | { "kind": "binary", operator: ArithmeticOperator, left: ArithmeticExpression, right: ArithmeticExpression, };