// apps/web/src/generated/wire/GenerateWorksheetRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Difficulty } from "./Difficulty";

export type GenerateWorksheetRequest = { schema_version: number, numeric_theme_id: number, seed: string, difficulty: Difficulty, timeout_ms: number | null, max_attempts: number | null, };